test(content-mapper): converge corpus/fixture tests on vitest + installed tsc
