+++ packages/content-mapper/README.md
@@ -1,7 +1,7 @@
 # Vue TypeScript content mapper
 
 `@vue/typescript-content-mapper` transforms Vue single-file components into the
-same TypeScript service scripts used by `vue-tsc`. It implements the TypeScript
+the TypeScript service scripts generated by `@vue/language-core`. It implements the TypeScript
 7 content-mapper protocol and keeps transforms parallel with a worker-thread
 pool.
 
@@ -49,19 +49,18 @@ TypeScript because Vue's generated helpers contain TypeScript syntax.
 Vue template `@vue-ignore` and `@vue-expect-error` comments are emitted through
 TypeScript content-mapper diagnostic directive mappings.
 
-## Full-corpus diagnostic parity
+## Full-corpus diagnostic snapshot
 
-Compare content-mapped tsgo output with `vue-tsc` across every project under
-`test-workspace/tsc`:
+Run the content mapper over every project under `test-workspace/tsc` and
+compare against the committed snapshot:
 
 ```sh
-pnpm test:content-mapper-parity -- --tsgo /absolute/path/to/tsgo
+pnpm test corpus
 ```
 
-The command generates temporary project sidecars, runs both compilers over the
-same 222-project graph, reports matching and implementation-specific output,
-then removes the sidecars. Add `--check` to fail on any difference, or
-`--output <directory>` to retain normalized outputs and a JSON report.
+The test generates temporary project sidecars, runs the installed native `tsc`
+with `--runExternalCode` over the 222-project graph, then snapshots the
+normalized diagnostics. Use `pnpm test corpus -- -u` to update the snapshot
+after an intentional change.
 
-See [MIGRATION.md](./MIGRATION.md) for current parity and language-server
-boundaries.
+See [MIGRATION.md](./MIGRATION.md) for known language-server boundaries.
