+++ packages/content-mapper/tests/fixtures.spec.ts
@@ -1,0 +1,22 @@
+import * as path from 'node:path';
+import { expect, test } from 'vitest';
+import { normalizeCompilerOutput, repositoryRoot, runTsc } from './utils';
+
+const configFileNames = [
+	'test-workspace/content-mapper/tsconfig.json',
+	'test-workspace/content-mapper-pug/tsconfig.json',
+	'test-workspace/content-mapper-directives/tsconfig.json',
+];
+
+test('content mapper fixtures', () => {
+	for (const configFileName of configFileNames) {
+		const result = runTsc([
+			'-p',
+			path.join(repositoryRoot, configFileName),
+			'--runExternalCode',
+			'--pretty',
+			'false',
+		]);
+		expect(normalizeCompilerOutput(result)).toMatchSnapshot(configFileName);
+	}
+});
